pyproject.toml
[project]
name = "molcrafts-molpy"
version = "0.1.0"
description = "Modern Python framework for molecular simulation and analysis"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "BSD-3-Clause"}

authors = [
    { name = "Roy Kid", email = "lijichen365@gmail.com" }
]

maintainers = [
    { name = "MolCrafts Team", email = "molcrafts@example.com" }
]

keywords = [
    "molecular-simulation",
    "computational-chemistry",
    "molecular-dynamics",
    "force-fields",
    "scientific-computing"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX',
    'Operating System :: Unix',
    'Operating System :: MacOS',
]

dependencies = [
    "molcrafts-molq",
    "numpy>=2.0",
    "python-igraph>=0.9.0",
    "lark>=1.1.0",
    "pint",
    "freud-analysis"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "pre-commit>=4.0.0"
]
all = [
    "molcrafts-molpy[dev]"
]
doc = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=8.0.0",
    "mkdocstrings[python]>=0.19.0"
]

[project.urls]
Homepage = "https://github.com/MolCrafts/molpy"
Documentation = "https://molcrafts.github.io/molpy"
Repository = "https://github.com/MolCrafts/molpy.git"
Issues = "https://github.com/MolCrafts/molpy/issues"
Changelog = "https://github.com/MolCrafts/molpy/blob/main/CHANGELOG.md"

[build-system]
requires = [
    "setuptools>=42",
    "wheel",
    "ninja",
    "numpy",
]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
